+++ Elements/src/Topography.cs
@@ -26,6 +26,8 @@ namespace Elements
 
         private double? _absoluteMinimumElevation;
 
+        internal List<Vertex> _baseVerts = new List<Vertex>();
+
         /// <summary>
         /// The maximum elevation of the topography.
         /// </summary>
@@ -168,14 +170,8 @@ namespace Elements
             {
                 if (args.PropertyName == "DepthBelowMinimumElevation" || args.PropertyName == "AbsoluteMinimumElevation")
                 {
-                    GenerateMeshAndSetInternals();
-                    absoluteMinimumElevation = this.AbsoluteMinimumElevation.HasValue ? this.AbsoluteMinimumElevation.Value : this.MinElevation - this.DepthBelowMinimumElevation;
-                    CreateSidesAndBottomMesh(this._mesh,
-                                             this.RowWidth,
-                                             absoluteMinimumElevation,
-                                             this.CellHeight,
-                                             this.CellWidth,
-                                             this.Origin);
+                    var minHeight = this.AbsoluteMinimumElevation.HasValue ? this.AbsoluteMinimumElevation.Value : this.MinElevation - this.DepthBelowMinimumElevation;
+                    MoveBaseVerts(minHeight);
                 }
             };
         }
@@ -204,6 +200,8 @@ namespace Elements
             var mesh = new Mesh();
             var triangles = (Math.Sqrt(elevations.Length) - 1) * rowWidth * 2;
 
+            var r = new Random();
+
             for (var y = 0; y < rowWidth; y++)
             {
                 for (var x = 0; x < rowWidth; x++)
@@ -226,7 +224,11 @@ namespace Elements
 
                     var uv = new UV(u, v);
 
-                    mesh.AddVertex(origin + new Vector3(x * cellWidth, y * cellHeight, el), uv: uv);
+                    // Add the tiniest amount of fuzz to avoid the 
+                    // triangles being identified as coplanar during
+                    // operations like CSG.
+                    var fuzz = r.NextDouble() * 0.0001;
+                    mesh.AddVertex(origin + new Vector3(x * cellWidth, y * cellHeight, el + fuzz), uv: uv);
 
                     if (y > 0 && x > 0)
                     {
@@ -251,7 +253,7 @@ namespace Elements
             return (mesh, maxElevation, minElevation);
         }
 
-        private static void CreateSidesAndBottomMesh(Mesh mesh,
+        private void CreateSidesAndBottomMesh(Mesh mesh,
                                                      int rowWidth,
                                                      double depth,
                                                      double cellHeight,
@@ -270,6 +272,8 @@ namespace Elements
             (Vector3 U, Vector3 V) basisTop = (default(Vector3), default(Vector3));
             (Vector3 U, Vector3 V) basisBottom = (default(Vector3), default(Vector3));
 
+            this._baseVerts.Clear();
+
             for (var u = 0; u < rowWidth - 1; u++)
             {
                 if (u == 0)
@@ -284,7 +288,7 @@ namespace Elements
                 Vertex l1 = null;
                 var i1 = u * rowWidth;
                 var v1Existing = mesh.Vertices[i1];
-                var v1 = mesh.AddVertex(v1Existing.Position);
+                var v1 = mesh.AddVertex(v1Existing.Position, normal: Vector3.XAxis.Negate());
                 if (lastL != null)
                 {
                     l1 = lastL;
@@ -293,6 +297,7 @@ namespace Elements
                 {
                     var p = new Vector3(v1.Position.X, v1.Position.Y, depth);
                     l1 = mesh.AddVertex(p);
+                    _baseVerts.Add(l1);
                 }
 
                 var i2 = i1 + rowWidth;
@@ -301,6 +306,7 @@ namespace Elements
 
                 var pl2 = new Vector3(v2.Position.X, v2.Position.Y, depth);
                 var l2 = mesh.AddVertex(pl2);
+                _baseVerts.Add(l2);
                 lastL = l2;
 
                 mesh.AddTriangle(l1, v1, v2);
@@ -310,7 +316,7 @@ namespace Elements
                 Vertex l3 = null;
                 var i3 = u * (rowWidth) + (rowWidth - 1);
                 var v3Existing = mesh.Vertices[i3];
-                var v3 = mesh.AddVertex(v3Existing.Position);
+                var v3 = mesh.AddVertex(v3Existing.Position, normal: Vector3.XAxis);
 
                 if (lastR != null)
                 {
@@ -320,6 +326,7 @@ namespace Elements
                 {
                     var p = new Vector3(v3.Position.X, v3.Position.Y, depth);
                     l3 = mesh.AddVertex(p);
+                    _baseVerts.Add(l3);
                 }
 
                 var i4 = i3 + rowWidth;
@@ -327,6 +334,7 @@ namespace Elements
                 var v4 = mesh.AddVertex(v4Existing.Position);
                 var pl4 = new Vector3(v4.Position.X, v4.Position.Y, depth);
                 var l4 = mesh.AddVertex(pl4);
+                _baseVerts.Add(l4);
                 lastR = l4;
 
                 mesh.AddTriangle(l3, v4, v3);
@@ -336,7 +344,7 @@ namespace Elements
                 Vertex l5 = null;
                 var i5 = u;
                 var v5Existing = mesh.Vertices[i5];
-                var v5 = mesh.AddVertex(v5Existing.Position);
+                var v5 = mesh.AddVertex(v5Existing.Position, normal: Vector3.YAxis);
                 if (lastT != null)
                 {
                     l5 = lastT;
@@ -345,6 +353,7 @@ namespace Elements
                 {
                     var p = new Vector3(v5.Position.X, v5.Position.Y, depth);
                     l5 = mesh.AddVertex(p);
+                    _baseVerts.Add(l5);
                 }
 
                 var i6 = i5 + 1;
@@ -352,6 +361,7 @@ namespace Elements
                 var v6 = mesh.AddVertex(v6Existing.Position);
                 var pl6 = new Vector3(v6.Position.X, v6.Position.Y, depth);
                 var l6 = mesh.AddVertex(pl6);
+                _baseVerts.Add(l6);
                 lastT = l6;
 
                 mesh.AddTriangle(l5, v6, v5);
@@ -361,7 +371,7 @@ namespace Elements
                 Vertex l7 = null;
                 var i7 = rowWidth * rowWidth - u - 1;
                 var v7Existing = mesh.Vertices[i7];
-                var v7 = mesh.AddVertex(v7Existing.Position);
+                var v7 = mesh.AddVertex(v7Existing.Position, normal: Vector3.YAxis.Negate());
 
                 if (lastB != null)
                 {
@@ -371,6 +381,7 @@ namespace Elements
                 {
                     var p = new Vector3(v7.Position.X, v7.Position.Y, depth);
                     l7 = mesh.AddVertex(p);
+                    _baseVerts.Add(l7);
                 }
 
                 var i8 = i7 - 1;
@@ -378,6 +389,7 @@ namespace Elements
                 var v8 = mesh.AddVertex(v8Existing.Position);
                 var pl8 = new Vector3(v8.Position.X, v8.Position.Y, depth);
                 var l8 = mesh.AddVertex(pl8);
+                _baseVerts.Add(l8);
                 lastB = l8;
 
                 mesh.AddTriangle(l7, v8, v7);
@@ -389,6 +401,8 @@ namespace Elements
             var bb2 = mesh.AddVertex(origin + new Vector3((rowWidth - 1) * cellWidth, 0, depth));
             var bb3 = mesh.AddVertex(origin + new Vector3((rowWidth - 1) * cellWidth, (rowWidth - 1) * cellHeight, depth));
             var bb4 = mesh.AddVertex(origin + new Vector3(0, (rowWidth - 1) * cellHeight, depth));
+
+            _baseVerts.AddRange(new[] { bb1, bb2, bb3, bb4 });
 
             mesh.AddTriangle(bb1, bb3, bb2);
             mesh.AddTriangle(bb1, bb4, bb3);
@@ -481,6 +495,154 @@ namespace Elements
                     return null;
             }
             return range.Select(i => this.Mesh.Vertices[start + i * increment]).ToArray();
+        }
+
+        /// <summary>
+        /// Trim the topography with the specified perimeter.
+        /// </summary>
+        /// <param name="perimeter">The perimeter of the trimmed topography.</param>
+        public void Trim(Polygon perimeter)
+        {
+            // This creates a tall trimming object which is extruded
+            // from the zero plane but then transformed in -Z by half its
+            // elevation to assure a trim through the object.
+
+            var topoCsg = this.Mesh.ToCsg();
+            var trim = new Extrude(perimeter, 200000, Vector3.ZAxis, false);
+            var t = new Transform(0, 0, -100000);
+            var trimCsg = trim._solid.ToCsg().Transform(t.ToMatrix4x4());
+            topoCsg = trimCsg.Intersect(topoCsg);
+            var mesh = new Mesh();
+            topoCsg.Tessellate(ref mesh);
+            this.Mesh = mesh;
+
+            // We've trimmed the mesh. We now need to
+            // reset the min and max elevation.
+            SetBaseVerts();
+            SetMinAndMaxElevation();
+        }
+
+        /// <summary>
+        /// Cut a tunnel through a topography.
+        /// </summary>
+        /// <param name="path">The path of the tunnel.</param>
+        /// <param name="diameter">The diameter of the tunnel.</param>
+        public void Tunnel(Curve path, double diameter)
+        {
+            if (diameter < 0.0)
+            {
+                throw new ArgumentException("Diameter must be greater than 0.0.", "diameter");
+            }
+
+            var topoCsg = this.Mesh.ToCsg();
+
+            var profile = new Circle(diameter / 2).ToPolygon(20);
+            var sweep = new Sweep(profile, path, 0.0, 0.0, 0.0, false);
+            var tunnelCsg = sweep._solid.ToCsg();
+
+            topoCsg = topoCsg.Substract(tunnelCsg);
+            var mesh = new Mesh();
+            topoCsg.Tessellate(ref mesh);
+            this.Mesh = mesh;
+
+            SetBaseVerts();
+            SetMinAndMaxElevation();
+        }
+
+        /// <summary>
+        /// Cut and or fill the topography with the specified perimeter to the specified elevation.
+        /// </summary>
+        /// <param name="perimeters">The perimeters of the cut and fill areas.</param>
+        /// <param name="elevation">The final elevation of the cut and fill.</param>
+        /// <param name="batterAngle">The angle of the battering surrounding the fill area in degrees.</param>
+        /// <param name="fills">Meshes representing the fill volume in the topography.</param>
+        /// <param name="cuts">Meshes representing the cut volume in the topography.</param>
+        /// <returns>The sum of all cut and fill volumes.</returns>
+        public (double CutVolume, double FillVolume) CutAndFill(IList<Polygon> perimeters, double elevation, out List<Mesh> cuts, out List<Mesh> fills, double batterAngle = 45.0)
+        {
+            if (this.AbsoluteMinimumElevation == null || elevation < this.AbsoluteMinimumElevation)
+            {
+                // Push the depth of the topography down past
+                // the required elevation.
+                this.AbsoluteMinimumElevation = elevation - 1;
+            }
+
+            if (batterAngle < 0.0)
+            {
+                throw new ArgumentException("The batter angle must be greater than 0.0", "batterAngle");
+            }
+
+            cuts = new List<Mesh>();
+            fills = new List<Mesh>();
+
+            var topoCsg = this.Mesh.ToCsg();
+            foreach (var p in perimeters)
+            {
+                topoCsg = CutAndFillInternal(topoCsg, elevation, p, cuts, fills, batterAngle);
+            }
+
+            var mesh = new Mesh();
+            topoCsg.Tessellate(ref mesh);
+            this.Mesh = mesh;
+
+            return (cuts.Sum(c => c.Volume()), fills.Sum(f => f.Volume()));
+        }
+
+        private Csg.Solid CutAndFillInternal(Csg.Solid topoCsg, double elevation, Polygon perimeter, List<Mesh> cuts, List<Mesh> fills, double batterAngle = 45.0)
+        {
+            // This check isn't perfect. Sites with topography may have
+            // areas where the elevation is lower, and areas where the elevation
+            // is higher.
+            if (elevation < this.MaxElevation)
+            {
+                // Cut
+                var cutSolid = new Extrude(perimeter, 100000, Vector3.ZAxis, false);
+                var t = new Transform(0, 0, elevation);
+                // Transform the cut volume to the elevation.
+                var cutCsg = cutSolid.Solid.ToCsg().Transform(t.ToMatrix4x4());
+
+                // Calculate the volume of the cut
+                // as the union of the two solids.
+                var cutXsect = topoCsg.Intersect(cutCsg);
+                var cut = new Mesh();
+                cutXsect.Tessellate(ref cut);
+                cuts.Add(cut);
+
+                topoCsg = topoCsg.Substract(cutCsg);
+            }
+
+            if (elevation > this.MinElevation)
+            {
+                // Fill
+                var height = elevation - this.MinElevation;
+
+                var fillSolid = new Extrude(perimeter, height, Vector3.ZAxis, false);
+                // Transform the fill volume down to the minimum elevation
+                // and fill up to the elevation.
+                var fillT = new Transform(0, 0, this.MinElevation);
+                var csgT = fillT.ToMatrix4x4();
+                var fillCsg = fillSolid.Solid.ToCsg().Transform(csgT);
+
+                var batterWidth = height * Math.Cos(Units.DegreesToRadians(batterAngle));
+                var batterProfile = new Polygon(new[]{
+                    Vector3.Origin,
+                    new Vector3(batterWidth, 0),
+                    new Vector3(0, height)
+                });
+
+                // Calculate the whole fill by adding the battering.
+                var batterSweep = new Sweep(batterProfile, perimeter, 0, 0, 0, false);
+                var batterCsg = batterSweep.Solid.ToCsg().Transform(csgT);
+                fillCsg = fillCsg.Union(batterCsg);
+
+                var xsect = fillCsg.Substract(topoCsg);
+                var fill = new Mesh();
+                xsect.Tessellate(ref fill);
+                fills.Add(fill);
+                topoCsg = topoCsg.Union(fillCsg);
+            }
+
+            return topoCsg;
         }
 
         /// <summary>
@@ -677,6 +839,50 @@ namespace Elements
             }
 
             return output;
+        }
+
+        private void MoveBaseVerts(double newElevation)
+        {
+            foreach (var v in this._baseVerts)
+            {
+                v.Position = new Vector3(v.Position.X, v.Position.Y, newElevation);
+            }
+        }
+
+        private void SetBaseVerts()
+        {
+            var minHeight = this.AbsoluteMinimumElevation.HasValue ? this.AbsoluteMinimumElevation.Value : this.MinElevation - this.DepthBelowMinimumElevation;
+            this._baseVerts.Clear();
+            foreach (var v in this.Mesh.Vertices)
+            {
+                if (v.Position.Z.ApproximatelyEquals(minHeight))
+                {
+                    this._baseVerts.Add(v);
+                }
+            }
+        }
+
+        private void SetMinAndMaxElevation()
+        {
+            // Ignore any vertices that have the depth of the sides.
+            var max = double.MinValue;
+            var min = double.MaxValue;
+            foreach (var v in this.Mesh.Vertices)
+            {
+                if (!_baseVerts.Contains(v))
+                {
+                    if (v.Position.Z > max)
+                    {
+                        max = v.Position.Z;
+                    }
+                    if (v.Position.Z < min)
+                    {
+                        min = v.Position.Z;
+                    }
+                }
+            }
+            this._minElevation = min;
+            this._maxElevation = max;
         }
 
         /// <summary>